+++ flask_monitoringdashboard/core/forms/daterange.py
@@ -5,6 +5,8 @@ from flask_wtf import FlaskForm
 from wtforms import validators, SubmitField
 from wtforms.fields.html5 import DateField
 
+from flask_monitoringdashboard.core.timezone import to_local_datetime
+
 DATE_FORMAT = '%Y-%m-%d'
 
 
@@ -13,7 +15,7 @@ class SelectDateRangeForm(FlaskForm):
     start_date = DateField('Start date', format=DATE_FORMAT, validators=[validators.data_required()])
     end_date = DateField('End date', format=DATE_FORMAT, validators=[validators.data_required()])
     submit = SubmitField('Submit')
-    type = 'SelectDateRangeForm'
+    title = 'Select two dates for reducing the size of the graph'
 
     def get_days(self):
         """
@@ -21,6 +23,21 @@ class SelectDateRangeForm(FlaskForm):
         """
         delta = self.end_date.data - self.start_date.data
         return [self.start_date.data + datetime.timedelta(days=i) for i in range(delta.days + 1)]
+
+    def content(self):
+        return '''
+          <div class="row">
+            <div class="col-sm-4"><i class="fa fa-calendar"></i> {} </div>
+            <div class="col-sm-4"><i class="fa fa-calendar"></i> {} </div>
+          </div>
+          <div class="row">
+            <div class="col-sm-4"> {} </div>
+            <div class="col-sm-4"> {} </div>
+            <div class="col-sm-4"> {} </div>
+          </div>'''.format(self.start_date.label, self.end_date.label,
+                           self.start_date(class_="form-control", required=True),
+                           self.end_date(class_="form-control", required=True),
+                           self.submit(class_="btn btn-primary btn-block"))
 
 
 def get_daterange_form(num_days=20):
@@ -36,6 +53,6 @@ def get_daterange_form(num_days=20):
         if form.start_date.data > form.end_date.data:
             form.start_date.data, form.end_date.data = form.end_date.data, form.start_date.data
     else:
-        form.end_date.data = datetime.date.today()
+        form.end_date.data = to_local_datetime(datetime.datetime.utcnow()).date()
         form.start_date.data = form.end_date.data - datetime.timedelta(days=num_days)
     return form