+++ flask_monitoringdashboard/core/utils.py
@@ -1,11 +1,10 @@
 import ast
-
 import numpy as np
 from flask import url_for
 from werkzeug.routing import BuildError
 
 from flask_monitoringdashboard import config
-from flask_monitoringdashboard.database.count import count_requests
+from flask_monitoringdashboard.database.count import count_requests, count_total_requests
 from flask_monitoringdashboard.database.endpoint import get_monitor_rule
 from flask_monitoringdashboard.database.function_calls import get_date_of_first_request
 
@@ -23,28 +22,17 @@ def get_endpoint_details(db_session, endpoint):
 def get_details(db_session):
     """ Return details about the deployment """
     import json
-    with open('flask_monitoringdashboard/constants.json', 'r') as f:
+    from flask_monitoringdashboard import loc
+    with open(loc() + 'constants.json', 'r') as f:
         constants = json.load(f)
 
     return {
         'link': config.link,
         'dashboard-version': constants['version'],
         'config-version': config.version,
-        'first-request': get_date_of_first_request(db_session)
+        'first-request': get_date_of_first_request(db_session),
+        'total-requests': count_total_requests(db_session)
     }
-
-
-def formatter(ms):
-    """
-    formats the ms into seconds and ms
-    :param ms: the number of ms
-    :return: a string representing the same amount, but now represented in seconds and ms.
-    """
-    sec = int(ms) // 1000
-    ms = int(ms) % 1000
-    if sec == 0:
-        return '{0}ms'.format(ms)
-    return '{0}.{1}s'.format(sec, ms)
 
 
 def get_url(end):