+++ flask_monitoringdashboard/database/endpoint.py
@@ -7,30 +7,43 @@ from sqlalchemy import func, desc
 from sqlalchemy.orm.exc import NoResultFound
 
 from flask_monitoringdashboard import config
+from flask_monitoringdashboard.core.timezone import to_local_datetime, to_utc_datetime
 from flask_monitoringdashboard.database import FunctionCall, MonitorRule
 
 
 def get_num_requests(db_session, endpoint, start_date, end_date):
     """ Returns a list with all dates on which an endpoint is accessed.
+        :param db_session: session containing the query
         :param endpoint: if None, the result is the sum of all endpoints
         :param start_date: datetime.date object
         :param end_date: datetime.date object
     """
-    query = db_session.query(func.strftime('%Y-%m-%d %H:00:00', FunctionCall.time).label('newTime'),
-                             func.count(FunctionCall.execution_time).label('count'))
+    query = db_session.query(FunctionCall.time)
     if endpoint:
         query = query.filter(FunctionCall.endpoint == endpoint)
-    result = query.filter(FunctionCall.time >= datetime.datetime.combine(start_date, datetime.time(0, 0, 0, 0))). \
-        filter(FunctionCall.time <= datetime.datetime.combine(end_date, datetime.time(23, 59, 59))). \
-        group_by('newTime').all()
-    db_session.expunge_all()
-    return result
+    result = query.filter(FunctionCall.time >= start_date, FunctionCall.time <= end_date).all()
+
+    return group_execution_times(result)
 
 
-def get_group_by_sorted(db_session, endpoint, limit=None):
+def group_execution_times(times):
+    """
+    Returns a list of tuples containing the number of hits per hour
+    :param times: list of datetime objects
+    :return: list of tuples ('%Y-%m-%d %H:00:00', count)
+    """
+    hours_dict = {}
+    for dt in times:
+        round_time = dt.time.strftime('%Y-%m-%d %H:00:00')
+        hours_dict[round_time] = hours_dict.get(round_time, 0) + 1
+    return hours_dict.items()
+
+
+def get_users(db_session, endpoint, limit=None):
     """
     Returns a list with the distinct group-by from a specific endpoint. The limit is used to filter the most used
-    distinct
+    distinct.
+    :param db_session: session containing the query
     :param endpoint: the endpoint to filter on
     :param limit: the number of
     :return: a list with the group_by as strings.
@@ -45,10 +58,11 @@ def get_group_by_sorted(db_session, endpoint, limit=None):
     return [r[0] for r in result]
 
 
-def get_ip_sorted(db_session, endpoint, limit=None):
+def get_ips(db_session, endpoint, limit=None):
     """
     Returns a list with the distinct group-by from a specific endpoint. The limit is used to filter the most used
-    distinct
+    distinct.
+    :param db_session: session containing the query
     :param endpoint: the endpoint to filter on
     :param limit: the number of
     :return: a list with the group_by as strings.
@@ -69,12 +83,13 @@ def get_monitor_rule(db_session, endpoint):
     try:
         result = db_session.query(MonitorRule). \
             filter(MonitorRule.endpoint == endpoint).one()
-        # for using the result when the session is closed, use expunge
-        db_session.expunge(result)
+        result.time_added = to_local_datetime(result.time_added)
+        result.last_accessed = to_local_datetime(result.last_accessed)
+        db_session.expunge_all()
         return result
     except NoResultFound:
         db_session.add(
-            MonitorRule(endpoint=endpoint, version_added=config.version, time_added=datetime.datetime.now()))
+            MonitorRule(endpoint=endpoint, version_added=config.version, time_added=datetime.datetime.utcnow()))
 
     # return new added row
     return get_monitor_rule(db_session, endpoint)
@@ -86,21 +101,10 @@ def update_monitor_rule(db_session, endpoint, value):
         update({MonitorRule.monitor: value})
 
 
-def get_all_measurement_per_column(db_session, endpoint, column, value):
-    """Return all entries with measurements from a given endpoint for which the column has a specific value.
-    Used for creating a box plot. """
-    result = db_session.query(FunctionCall).filter(FunctionCall.endpoint == endpoint, column == value).all()
-    db_session.expunge_all()
-    return result
-
-
-def get_last_accessed_times(db_session, endpoint):
-    """ Returns a list of all endpoints and their last accessed time. """
-    result = db_session.query(MonitorRule.last_accessed).filter(MonitorRule.endpoint == endpoint).first()
-    db_session.expunge_all()
-    if result:
-        return result[0]
-    return None
+def get_last_accessed_times(db_session):
+    """ Returns the accessed time of a single endpoint. """
+    result = db_session.query(MonitorRule.endpoint, MonitorRule.last_accessed).all()
+    return [(end, to_local_datetime(time)) for end, time in result]
 
 
 def update_last_accessed(db_session, endpoint, value):