+++ flask_monitoringdashboard/test/views/test_details.py
@@ -16,34 +16,34 @@ class TestResult(unittest.TestCase):
         """
             Just retrieve the content and check if nothing breaks
         """
-        test_admin_secure(self, 'result/{}/heatmap'.format(NAME))
+        test_admin_secure(self, 'endpoint/{}/hourly_load'.format(NAME))
 
     def test_result_time_per_version_per_user(self):
         """
             Just retrieve the content and check if nothing breaks
         """
-        test_admin_secure(self, 'result/{}/time_per_version_per_user'.format(NAME))
+        test_admin_secure(self, 'endpoint/{}/version_user'.format(NAME))
 
     def test_result_time_per_version_per_ip(self):
         """
             Just retrieve the content and check if nothing breaks
         """
-        test_admin_secure(self, 'result/{}/time_per_version_per_ip'.format(NAME))
+        test_admin_secure(self, 'endpoint/{}/version_ip'.format(NAME))
 
     def test_result_time_per_version(self):
         """
             Just retrieve the content and check if nothing breaks
         """
-        test_admin_secure(self, 'result/{}/time_per_version'.format(NAME))
+        test_admin_secure(self, 'endpoint/{}/versions'.format(NAME))
 
     def test_result_time_per_user(self):
         """
             Just retrieve the content and check if nothing breaks
         """
-        test_admin_secure(self, 'result/{}/time_per_user'.format(NAME))
+        test_admin_secure(self, 'endpoint/{}/users'.format(NAME))
 
     def test_result_outliers(self):
         """
             Just retrieve the content and check if nothing breaks
         """
-        test_admin_secure(self, 'result/{}/outliers'.format(NAME))
+        test_admin_secure(self, 'endpoint/{}/outliers'.format(NAME))